# AI Generated Gnome API - OpenAPI 3.0.yaml
openapi: 3.0.3
info:
  title: Gnome Board Game Shop API
  description: |
    RESTful API for managing board game products, categories, variants, and admin authentication.
    
    ## Authentication
    This API uses JWT Bearer token authentication for admin operations. 
    - Public endpoints (GET) don't require authentication
    - Admin endpoints (POST, PUT, DELETE) require valid JWT token with Admin role
    
    ## Getting Started
    1. Use the `/auth/login` endpoint to authenticate and get JWT token
    2. Include the JWT token in the Authorization header for protected endpoints
    3. Use `/auth/refresh` to get new tokens when the current one expires
    
    ## Rate Limiting
    - 100 requests per minute per IP address
    - 1000 requests per hour per authenticated user
    
    ## Error Handling
    All errors return a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: Gnome Board Game Shop
    email: support@gnomeboardgames.com
    url: https://gnomeboardgames.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:7080
    description: Development server
  - url: https://api.gnomeboardgames.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /category/list:
    get:
      tags:
        - Categories
      summary: Get categories list
      description: Retrieves a paginated list of categories with optional filtering and sorting
      operationId: listCategories
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page-size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: date-from
          in: query
          description: Filter categories created from this date (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: date-to
          in: query
          description: Filter categories created until this date (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: name
          in: query
          description: Filter categories by name (partial match)
          required: false
          schema:
            type: string
        - name: sort-by
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [created-date-time, name, updated-date-time]
            default: created-date-time
        - name: sort-order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /category/{slug}:
    get:
      tags:
        - Categories
      summary: Get category by slug
      description: Retrieves detailed information about a specific category with products count using its slug
      operationId: getCategoryBySlug
      parameters:
        - name: slug
          in: path
          description: The unique slug identifier of the category
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Categories
      summary: Delete category
      description: Permanently deletes a category. Requires admin authentication.
      operationId: deleteCategory
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: The unique identifier of the category to delete
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDeletedResponse'
        '401':
          description: Unauthorized - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /category/add:
    post:
      tags:
        - Categories
      summary: Add new category
      description: Creates a new category. Requires admin authentication.
      operationId: addCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AddCategoryRequest'
      responses:
        '200':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryCreatedResponse'
        '400':
          description: Invalid category data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /category/update:
    put:
      tags:
        - Categories
      summary: Update category
      description: Updates an existing category. Requires admin authentication.
      operationId: updateCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryUpdatedResponse'
        '400':
          description: Invalid category data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme. Enter your token in the text input below.
  schemas:
...
    CategoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        totalCount:
          type: integer
          description: Total number of categories
          example: 1
        pageNumber:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 10
        totalPages:
          type: integer
          description: Total number of pages
          example: 1
        hasPreviousPage:
          type: boolean
          description: Whether there is a previous page
          example: false
        hasNextPage:
          type: boolean
          description: Whether there is a next page
          example: false
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          description: Category ID
          example: 1
        name:
          type: string
          description: Category name
          example: "Strategy Games"
        slug:
          type: string
          description: URL-friendly category name
          example: "strategy-games"
        productsCount:
          type: integer
          description: Number of products in this category
          example: 5
        createdDateTime:
          type: string
          format: date-time
          description: Category creation date
          example: "2024-01-01T00:00:00Z"
        updatedDateTime:
          type: string
          format: date-time
          description: Category last update date
          example: "2024-01-15T10:00:00Z"
...
    AddCategoryRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          description: Category name
          example: "Strategy Games"
        slug:
          type: string
          description: URL-friendly category name
          example: "strategy-games"
    UpdateCategoryRequest:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: integer
          description: Category ID to update
          example: 1
        name:
          type: string
          description: Updated category name
          example: "Updated Strategy Games"
        slug:
          type: string
          description: Updated URL-friendly category name
          example: "updated-strategy-games"
    CategoryCreatedResponse:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the newly created category
          example: 2
    CategoryUpdatedResponse:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the updated category
          example: 1
    CategoryDeletedResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Confirmation that the category was deleted successfully
          example: true

    # Variant Schemas
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message describing what went wrong
          example: "Invalid username or password" 